Code review feedback + bug fixes
- Fixes bug 1750200 (yaml_set_anchor issue)
