+++ setup.cfg
@@ -8,7 +8,7 @@ version = 0.0.1.a0
 author = Zeid Zabaneh
 author_email = zeid@mozilla.com
 description =  Module Ownership in Tree System
-long_description = file: README.md
+long_description = "Full documentation can be found at https://mots.readthedocs.io/en/develop/."
 long_description_content_type = text/markdown
 url = https://github.com/mozilla-conduit/mots
 classifiers =
