+++ src/mots/module.py
@@ -5,19 +5,10 @@
 """Module operations and utility functions."""
 
 from __future__ import annotations
-from collections import defaultdict
 import logging
 from pathlib import Path
-from mots.config import FileConfig
-from mots.utils import generate_machine_readable_name
 
 logger = logging.getLogger(__name__)
-
-
-class ValidationError(TypeError):
-    """Thrown when a particular module is not valid."""
-
-    pass
 
 
 class Module:
@@ -201,140 +192,3 @@ def show(modules: list[Module], module: str):
             print(_module)
             return
     raise ValueError(f"{module} not found.")
-
-
-def clean(file_config: FileConfig, write: bool = True):
-    """Clean and re-sort configuration file.
-
-    Load condiguration from disk, sort modules and submodules by `machine_name`. If
-    there is no valid `machine_name`, generate one. Write changes to disk.
-
-    :param file_config: an instance of :class:`FileConfig`
-    :param write: if set to `True`, writes changes to disk.
-    """
-    from mots.directory import Directory
-
-    file_config.load()
-    directory = Directory(file_config)
-    directory.load()
-    for i in range(len(file_config.config["modules"])):
-        module = file_config.config["modules"][i]
-        if "machine_name" not in module:
-            module["machine_name"] = generate_machine_readable_name(module["name"])
-
-        people_keys = ("owners", "peers")
-        for key in people_keys:
-            if key in module and module[key]:
-                for i in range(len(module[key])):
-                    person = module[key][i]
-                    module[key][i] = file_config.config["people"][
-                        directory.people.by_bmo_id[person["bmo_id"]]
-                    ]
-
-        # Do the same for submodules.
-        if "submodules" in module and module["submodules"]:
-            module["submodules"].sort(key=lambda x: x["name"])
-            for submodule in module["submodules"]:
-                for key in people_keys:
-                    if key in submodule and submodule[key]:
-                        for i in range(len(submodule[key])):
-                            person = submodule[key][i]
-                            submodule[key][i] = file_config.config["people"][
-                                directory.people.by_bmo_id[person["bmo_id"]]
-                            ]
-                if "machine_name" not in submodule:
-                    submodule["machine_name"] = generate_machine_readable_name(
-                        submodule["name"]
-                    )
-
-    file_config.config["modules"].sort(key=lambda x: x["machine_name"])
-
-    nicks = []
-    for p in file_config.config["people"]:
-        machine_readable_nick = generate_machine_readable_name(
-            p["nick"], keep_case=True
-        )
-        if machine_readable_nick in nicks or not machine_readable_nick:
-            continue
-        nicks.append(machine_readable_nick)
-        try:
-            p.yaml_set_anchor(machine_readable_nick)
-        except Exception as e:
-            # NOTE: this happens when the file changes and we lose ruamel embedded data
-            # TODO: fix this
-            logger.error(e)
-
-    if write:
-        file_config.write()
-
-
-def validate(config: dict, repo_path: str):
-    """Validate the current state of the config file.
-
-    - Check if top-level dictionary contains required keys
-    - Check if machine names are unique
-    - Instantiate and run validation on each :class:`Module` instance
-
-    :raises ValidationError: if any validation errors are detected
-    """
-    # Validate that config has all the required keys.
-    required_keys = ("repo", "created_at", "updated_at", "modules")
-    keys_diff = required_keys - config.keys()
-    if len(keys_diff) != 0:
-        raise ValidationError(f"{keys_diff} missing from configuration file.")
-
-    modules = config["modules"]
-
-    # Count machine name repetitions in a defaultdict.
-    machine_names = defaultdict(int)
-    for m in modules:
-        machine_names[m["machine_name"]] += 1
-        if "submodules" in m and m["submodules"]:
-            for sm in m["submodules"]:
-                machine_names[sm["machine_name"]] += 1
-
-    machine_names = {name: count for name, count in machine_names.items() if count > 1}
-    if machine_names:
-        raise ValidationError(
-            f"Duplicate machine name(s) found: {', '.join(machine_names.keys())}"
-        )
-
-    errors = []
-    for i in range(len(config["modules"])):
-        module = config["modules"][i]
-        module = Module(repo_path=repo_path, **module)
-        validation_errors = module.validate()
-        if validation_errors:
-            errors.append(validation_errors)
-
-    if errors:
-        raise ValidationError(errors)
-    logger.info("All modules validated successfully.")
-
-
-def add(
-    new_module: dict, file_config: FileConfig, parent: str = None, write: bool = True
-):
-    """Add a new module to the configuration.
-
-    :param module: a dictionary containing module parameters
-    :param file_config: an instance of :class:`FileConfig`
-    :param parent: the machine name of the parent module if applicable
-    :param write: if set to `True`, writes changes to disk
-    """
-    file_config.load()
-    modules = file_config.config["modules"]
-    serialized = Module(**new_module, repo_path=file_config.repo_path).serialize()
-
-    if parent:
-        for module in modules:
-            if module["machine_name"] == parent:
-                if "submodules" not in module or not module["submodules"]:
-                    module["submodules"] = []
-                module["submodules"].append(serialized)
-                break
-    else:
-        modules.append(serialized)
-
-    if write:
-        file_config.write()
